refactor: modularize zustand store with slices
